requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
ffprobe==0.5
future==0.18.2
gTTS==2.1.0
gTTS-token==1.1.3
idna==2.8
pycparser==2.19
pydub==0.23.1
python-telegram-bot==12.2.0
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
tornado==6.0.3
urllib3==1.25.7
